+++ aiohomematic/model/custom/cover.py
@@ -9,7 +9,7 @@ Public API of this module is defined by __all__.
 from __future__ import annotations
 
 import asyncio
-from enum import IntEnum, StrEnum
+from enum import IntEnum, StrEnum, unique
 import logging
 from typing import Final, Unpack
 
@@ -40,6 +40,7 @@ _OPEN_TILT_LEVEL: Final = 1.0
 _WD_CLOSED_LEVEL: Final = -0.005
 
 
+@unique
 class _CoverActivity(StrEnum):
     """Enum with cover activities."""
 
@@ -47,6 +48,7 @@ class _CoverActivity(StrEnum):
     OPENING = "UP"
 
 
+@unique
 class _CoverPosition(IntEnum):
     """Enum with cover positions."""
 
@@ -55,6 +57,7 @@ class _CoverPosition(IntEnum):
     CLOSED = 0
 
 
+@unique
 class _GarageDoorActivity(IntEnum):
     """Enum with garage door commands."""
 
@@ -62,6 +65,7 @@ class _GarageDoorActivity(IntEnum):
     OPENING = 2
 
 
+@unique
 class _GarageDoorCommand(StrEnum):
     """Enum with garage door commands."""
 
@@ -72,6 +76,7 @@ class _GarageDoorCommand(StrEnum):
     STOP = "STOP"
 
 
+@unique
 class _GarageDoorState(StrEnum):
     """Enum with garage door states."""
 
@@ -81,6 +86,7 @@ class _GarageDoorState(StrEnum):
     POSITION_UNKNOWN = "_POSITION_UNKNOWN"
 
 
+@unique
 class _StateChangeArg(StrEnum):
     """Enum with cover state change arguments."""